refactor: Dropdown 로직을 훅으로 분리
